requirements.txt
pandas
openpyxl
openai
python-dotenv
chromadb
sentence-transformers
tqdm
